updated repository for contacts

+++ src/repository/contacts.py
@@ -6,41 +6,40 @@ from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import or_, and_, extract
 
-from src.entity.models import Contact
+from src.entity.models import Contact, User
+from src.schema import contact
 from src.schema.contact import ContactSchema, ContactUpdate
 
 
-async def get_contacts(limit: int, offset: int, name: str, surname: str, email: str, db: AsyncSession):
-    query = select(Contact).offset(offset).limit(limit)
+async def get_contacts(limit: int, offset: int, name: str, surname: str, email: str, db: AsyncSession,
+                       current_user: User):
+    statement = select(Contact).filter_by(user=current_user).offset(offset).limit(limit)
 
-    filters = []
     if name:
-        filters.append(Contact.name.ilike(f"%{name}%"))
+        statement = statement.filter(Contact.name.ilike(f"%{name}%"))
     if surname:
-        filters.append(Contact.surname.ilike(f"%{surname}%"))
+        statement = statement.filter(Contact.surname.ilike(f"%{surname}%"))
     if email:
-        filters.append(Contact.email.ilike(f"%{email}%"))
+        statement = statement.filter(Contact.email.ilike(f"%{email}%"))
 
-    if filters:
-        query = query.where(and_(*filters))
-
-    contacts = await db.execute(query)
+    contacts = await db.execute(statement)
+    await db.close()
     return contacts.scalars().all()
 
 
-async def get_contact(contact_id: int, db: AsyncSession):
-    statement = select(Contact).filter_by(id=contact_id)
+async def get_contact(contact_id: int, db: AsyncSession, current_user: User):
+    statement = select(Contact).filter_by(id=contact_id, user=current_user)
     contact = await db.execute(statement)
     await db.close()
     return contact.scalar_one_or_none()
 
 
-async def create_contact(body: ContactSchema, db: AsyncSession):
+async def create_contact(body: ContactSchema, db: AsyncSession, current_user: User):
     if body.birthday >= date.today():
         raise HTTPException(status_code=400, detail="Birthday must be in the past")
 
     try:
-        contact = Contact(**body.model_dump())
+        contact = Contact(**body.model_dump(), user=current_user)
         db.add(contact)
         await db.commit()
         await db.refresh(contact)
@@ -52,43 +51,38 @@ async def create_contact(body: ContactSchema, db: AsyncSession):
         await db.close()
 
 
-async def update_contact(contact_id: int, contact_update: ContactUpdate, db: AsyncSession):
-    query = select(Contact).filter_by(id=contact_id)
-    result = await db.execute(query)
-    existing_contact = result.scalar_one_or_none()
+async def update_contact(contact_id: int, contact_update: ContactUpdate, db: AsyncSession, current_user: User):
+    statement = select(Contact).filter_by(id=contact_id, user=current_user)
+    existing_contact = await db.execute(statement)
+    existing_contact = existing_contact.scalar_one_or_none()
 
     if not existing_contact:
         raise HTTPException(status_code=404, detail="Contact not found")
 
-    for key, value in contact_update.model_dump(exclude_unset=True).items():
-        setattr(existing_contact, key, value)
-
-    try:
+    if existing_contact:
+        for key, value in contact.model_dump().items():
+            setattr(existing_contact, key, value)
         await db.commit()
         await db.refresh(existing_contact)
-        return existing_contact
-    except SQLAlchemyError as e:
-        await db.rollback()
-        raise HTTPException(status_code=500, detail=str(e))
+        await db.close()
+    return existing_contact
 
 
-async def delete_contact(contact_id: int, db: AsyncSession):
-    query = select(Contact).filter_by(id=contact_id)
-    result = await db.execute(query)
-    contact = result.scalar_one_or_none()
+async def delete_contact(contact_id: int, db: AsyncSession, current_user: User):
+    statement = select(Contact).filter_by(id=contact_id, user=current_user)
+    contact = await db.execute(statement)
+    contact = contact.scalar_one_or_none()
 
     if not contact:
         raise HTTPException(status_code=404, detail="Contact not found")
 
-    await db.delete(contact)
-    try:
+    if contact:
+        await db.delete(contact)
         await db.commit()
-    except SQLAlchemyError as e:
-        await db.rollback()
-        raise HTTPException(status_code=500, detail=str(e))
+        return contact
 
 
-async def get_upcoming_birthdays(db: AsyncSession):
+async def get_upcoming_birthdays(db: AsyncSession, current_user: User):
     today = date.today()
     next_week = today + timedelta(days=7)
 
@@ -117,4 +111,5 @@ async def get_upcoming_birthdays(db: AsyncSession):
     )
 
     result = await db.execute(query)
+    await db.close()
     return result.scalars().all()
